Fix two pre-existing branching test failures
1. test_field_type_change_text_to_integer_merge: update expected PostgreSQL
   column type from integer to bigint. CO integer fields use BigIntegerField
   so alter_field produces a bigint column, not integer.

2. test_self_referential_multiobject_field_merge_and_revert: fix squash
   dep-graph building for dynamically-generated CO models. ContentType.
   model_class() returns None for CO objects during the squash dep-graph
   phase (the COT CREATE has not been applied to main yet), so the
   meta-is-None guard was silently skipping all CO object changes and no
   M2M dep edges were injected. Fix: detect CO models via cc.key[0] (the
   app_label.model_name string, always available) and add a fallback in
   _collect_co_refs that scans postchange_data for list-of-int values (the
   only CO field type that produces these is a direct M2M) to add
   self-referential dep edges even when the model class is not loaded.

+++ netbox_custom_objects/tests/test_branching.py
@@ -2504,7 +2504,7 @@ class FieldAttributeChangesTestCase(BranchingTestBase, _TestBase):
         field_main = CustomObjectTypeField.objects.get(pk=field_pk)
         self.assertEqual(field_main.type, 'integer')
 
-        # PostgreSQL column type must be integer.
+        # PostgreSQL column type must be bigint (CO integer fields use BigIntegerField).
         cot_main = CustomObjectType.objects.get(pk=cot_pk)
         co_table = cot_main.get_database_table_name()
         with main_conn.cursor() as cursor:
@@ -2514,7 +2514,7 @@ class FieldAttributeChangesTestCase(BranchingTestBase, _TestBase):
                 [co_table, 'value'],
             )
             data_type = cursor.fetchone()[0]
-        self.assertEqual(data_type, 'integer')
+        self.assertEqual(data_type, 'bigint')
 
         # CO value survived the cast.
         co_main = cot_main.get_model().objects.get(pk=co_pk)
